requirements.txt
discord.py
capmonster_python
pyyaml
httpx
py-cord